requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.1.1
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
SQLAlchemy==2.0.21
# psycopg2-binary==2.9.7  # Non necessario per SQLite
Werkzeug==2.3.7
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
requests==2.31.0
joblib==1.3.2
